refactor(train): fold the method shell into cordis_backend
The recipe and processor bind one to one with the engine: the method
variability of harness evolution lives in the user's propose and
evaluate callables, so the shell is part of the engine's public
surface. This mirrors slime_backend, which carries its own bridge and
driver glue inside the package. The short lived reef/harness_evolve
top level package is gone; imports and the implementation class path
move to reef.train.cordis_backend.
