+++ .github/CODEOWNERS
@@ -19,7 +19,7 @@
 /recipes/sao/ @Benjamin-eecs
 /recipes/tttd/ @Benjamin-eecs
 /recipes/openclawrl/ @Benjamin-eecs
-/reef/harness_evolve/ @Benjamin-eecs
+/reef/train/cordis_backend/ @Benjamin-eecs
 /reef/surface/ @Benjamin-eecs
 /third_party/cordis/ @Benjamin-eecs
 /tutorials/harness_evolve/ @Benjamin-eecs
