+++ .github/labeler.yml
@@ -30,7 +30,7 @@
           - "recipes/sao/**"
           - "recipes/tttd/**"
           - "recipes/openclawrl/**"
-          - "reef/harness_evolve/**"
+          - "reef/train/cordis_backend/**"
           - "reef/surface/**"
           - "third_party/cordis/**"
           - "tutorials/harness_evolve/**"
