+++ tests/reef_service/test_harness_recipe.py
@@ -18,12 +18,17 @@ from reef.dispatcher import Dispatcher
 from reef.harness.adapters import get_adapter
 from reef.harness.episode import EpisodeResult
 from reef.harness.model_binding import ModelBinding, ModelBindingError, ModelBindings
-from reef.harness_evolve import HarnessEvolveRecipe
 from reef.recipe import RecipeConfigError
 from reef.recipe.registry import RecipeRegistry, recipe_class_for
 from reef.records import RecordStore
 from reef.runtime.adapters.inference_proxy import InferenceProxyRuntime
-from reef.train.cordis_backend import CordisBackend, Mutation, MutationError, ScoreComparisonSelector
+from reef.train.cordis_backend import (
+    CordisBackend,
+    HarnessEvolveRecipe,
+    Mutation,
+    MutationError,
+    ScoreComparisonSelector,
+)
 from reef.train.cordis_backend.strategies import resolve_episode_scorer, resolve_proposer
 from reef.train.evaluation import DefaultCandidateEvaluationPlugin
 from reef.train.trainer import Trainer
@@ -148,7 +153,7 @@ def run_backend_step(
 
 
 def test_recipe_resolves_by_dotted_reference() -> None:
-    assert recipe_class_for("reef.harness_evolve.recipe:HarnessEvolveRecipe") is HarnessEvolveRecipe
+    assert recipe_class_for("reef.train.cordis_backend.recipe:HarnessEvolveRecipe") is HarnessEvolveRecipe
     assert recipe_class_for("harness_evolve") is None
