+++ tests/reef_service/test_skillclaw.py
@@ -18,11 +18,11 @@ import yaml
 
 from reef.core import AgentRecord, RequestType
 from reef.harness.model_binding import ModelBinding, ModelBindings
-from reef.harness_evolve.processor import HarnessEvolveProcessor
 from reef.recipe import RecipeConfigError
 from reef.recipe.registry import build_recipe
 from reef.runtime.adapters.inference_proxy import InferenceProxyRuntime
 from reef.train.cordis_backend import Mutation
+from reef.train.cordis_backend.processor import HarnessEvolveProcessor
 from reef.train.cordis_backend.strategies import resolve_episode_scorer, resolve_proposer
 from reef.train.evaluation import AlwaysSelect
 from reef.train.types import ProcessorContext, TraceSample
